C++ Dockerfile: Remove pin of libc version number (#132)
Previously we pinned libc6 to version 2.31. This no longer works, with the following error on `apt install`:
```
E: Version '2.31-0ubuntu9.9' for 'libc6' was not found
E: Version '2.31-0ubuntu9.9' for 'libc6-dev' was not found
E: Version '2.31-0ubuntu9.9' for 'libc-dev-bin' was not found
```
I removed the pin, rebuilt the image, and checked the `libc` version installed via `ldd --version`. The version is `2.31-0ubuntu9.12`, so I think we're safe to stop forcing the version.