Add config param to force GTP to not allow rule changes (#139)
Our bots are only trained with Tromp-Taylor rules, but KGS doesn't have Tromp-Taylor rules. If a bot plays with non-Tromp-Taylor rules, then because the rules are features of the NN input, all NN inputs will be out-of-distribution. We therefore want to force our bots to always think they're playing with Tromp-Taylor rules (at the very least for NN evaluation)

This PR:
* Incorporates https://github.com/AlignmentResearch/KataGo-custom/pull/109: adds a config param ignoreGTPRuleChanges (in the same vein as the existing KataGo-raw config param ignoreGTPAndForceKomi) that makes KataGo ignore rule changes from `katago-set-rule`, `katago-set-rules`, `kgs-rules`, and `loadsgf`.
* Unrelatedly: default `multiStoneSuicideLegal = true` in our GTP config (previously we had made this false for Leela/ELF transfer experiments, but these days all our GTP evals are not against Leela/ELF and I don't want to make the mistake of forgetting to set this param to `true`)